aws/zillow/requirements.txt
requests
lxml
boto3
psycopg2-binary
